Validate file name on four or five digits in the part of the file name before the underscore.
e.g from [0-9]{4} to [0-9]{4,5}

Apparently at some point with v.29 or 30 of MythTV,
they've changed the file naming convention for recordings
and Myth2Kodi was no longer recognizing legit. recordings
as being MythTV recordings. The name of the recording files
are now being prefixed by a digit that represents
(what I guess to be) the recording back-end.

I've seen this in a few installations, but haven't confirmed
the reason for the naming change.

For example: What used to be named "1302_2020....ts"
is now being called "11302_2020....ts" by Myth.

To keep things working I made an edit to the regex
at about line #860:

FROM:
RECPATTERN='[0-9]{4}_20[0-9]{2}[0-1][0-9][0-3][0-9][0-2][0-9][0-5][0-9]{3}\.'

TO:
RECPATTERN='[0-9]{4,5}_20[0-9]{2}[0-1][0-9][0-3][0-9][0-2][0-9][0-5][0-9]{3}\.'